Add initial feedback as part of the Discovery Details Page
Note: a slightly different approach is applied here because the code
itself for Discovery is not yet on GitHub. The 'old' peer review template
is used and - because a number of changes have been identified - the
pre-release accessibility tests will be performed once those issues
identified as a priority have been addressed.